requirements.txt
rasa_nlu==0.11.4
rasa_core==0.8.5
scikit-learn==0.19.1
scipy==1.0.1
pymongo==3.6.1
numpy==1.14.2
spacy==2.0.11
sklearn_crfsuite==0.3.6
falcon==1.4.1
gunicorn==19.7.1
gevent==1.2.2